I modified compute_msms_buried_surface_area.py so that it now splits interface files by connected components. In addition the surface area of the connected components must now be less than X (by default, all interfaces less than 100 angstroms are filtered).
I also modified where files are output to output/db/PDBI/, where
db are the two middle letters of the PDBid and PDBI is the PDBid.

I added the script master_ppi.sh which receives as input a PDB id,
mounts Jaume's PDB database and runs compute_msms_buried_surface_area.py
on it.

Finally, I modified loadMSMS.py so that it now doesn't crash when faces
are not present.